del fonder oraxen and check specific_properties

+++ run.py
@@ -22,6 +22,8 @@ def rename_keys(dict_, new_keys):
 if not os.path.isdir('./ItemsAdder'):
     os.mkdir('./ItemsAdder') 
     os.mkdir('./ItemsAdder/data') 
+if os.path.isdir('./Oraxen'):
+    shutil.rmtree('./Oraxen')
 if not os.path.isdir('./Oraxen'):
     os.mkdir('./Oraxen') 
     os.mkdir('./Oraxen/items') 
@@ -140,44 +142,45 @@ for get_namespace in os.listdir(itemadder):
                         documents['items'][key]['Pack']['model'] = get_namespace+"/"+documents['items'][key]['Pack']['model']
                         
                     else:
-                        if 'armor' in documents['items'][key]['specific_properties']:
-                            documents['items'][key]['Pack']['parent_model']= "item/generated"
-                            a_text = get_namespace+"/"+documents['items'][key]['Pack']['textures'][0]
-                            documents['items'][key]['Pack']['textures'] = [a_text,a_text]
+                        if 'specific_properties' in documents['items'][key]:
+                            if 'armor' in documents['items'][key]['specific_properties']:
+                                documents['items'][key]['Pack']['parent_model']= "item/generated"
+                                a_text = get_namespace+"/"+documents['items'][key]['Pack']['textures'][0]
+                                documents['items'][key]['Pack']['textures'] = [a_text,a_text]
 
-                            list_type_arror = {"chest":"chestplate","legs":"leggings","feet":"boots"}
-                            
-                            documents['items'][key]['material'] = "LEATHER_"+list_type_arror[documents['items'][key]['specific_properties']['armor']['slot']].upper()
-               
-                            vv = documents['items'].pop(key)
-                            nv = a_text.split("/")[2].split("_")[0]
-                            nk = nv+'_'+list_type_arror[vv['specific_properties']['armor']['slot']]
-                            documents['items'][nk] = vv
-                            documents['items'][nk].pop('specific_properties')
-                            
-                            
+                                list_type_arror = {"chest":"chestplate","legs":"leggings","feet":"boots"}
+                                
+                                documents['items'][key]['material'] = "LEATHER_"+list_type_arror[documents['items'][key]['specific_properties']['armor']['slot']].upper()
+                
+                                vv = documents['items'].pop(key)
+                                nv = a_text.split("/")[2].split("_")[0]
+                                nk = nv+'_'+list_type_arror[vv['specific_properties']['armor']['slot']]
+                                documents['items'][nk] = vv
+                                documents['items'][nk].pop('specific_properties')
+                                
+                                
 
-                            old_file = ''
-                            new_file = ''
-                            # replace name armor
-                            if os.path.exists(f"Oraxen/pack/textures/{get_namespace}/armor/{nv}_chestplate.png"):
-                                old_file = os.path.join(f"Oraxen/pack/textures/{get_namespace}/armor", f"{nv}_chestplate.png")
-                                new_file = os.path.join(f"Oraxen/pack/textures/{get_namespace}/armor", f"{nv}_armor_layer_1.png")
-                                if not os.path.exists(f"Oraxen/pack/textures/{get_namespace}/armor/{nv}_armor_layer_1.png"):
-                                    os.rename(old_file, new_file)
-                                if os.path.exists(old_file):
-                                    os.remove(old_file)
-                            if os.path.exists(f"Oraxen/pack/textures/{get_namespace}/armor/{nv}_leggings.png"):
-                                old_file_2 = os.path.join(f"Oraxen/pack/textures/{get_namespace}/armor", f"{nv}_leggings.png")
-                                new_file_2 = os.path.join(f"Oraxen/pack/textures/{get_namespace}/armor", f"{nv}_armor_layer_2.png")
-                                if not os.path.exists(f"Oraxen/pack/textures/{get_namespace}/armor/{nv}_armor_layer_2.png"):
-                                    os.rename(old_file_2, new_file_2)
-                                if os.path.exists(old_file_2):
-                                    os.remove(old_file_2)
-                                                            
-                            # set armor 128x32 config
-                            if not os.path.exists(f"Oraxen/settings.yml"):
-                                shutil.copy(f"Oraxen_settings.yml", f"Oraxen/settings.yml")
+                                old_file = ''
+                                new_file = ''
+                                # replace name armor
+                                if os.path.exists(f"Oraxen/pack/textures/{get_namespace}/armor/{nv}_chestplate.png"):
+                                    old_file = os.path.join(f"Oraxen/pack/textures/{get_namespace}/armor", f"{nv}_chestplate.png")
+                                    new_file = os.path.join(f"Oraxen/pack/textures/{get_namespace}/armor", f"{nv}_armor_layer_1.png")
+                                    if not os.path.exists(f"Oraxen/pack/textures/{get_namespace}/armor/{nv}_armor_layer_1.png"):
+                                        os.rename(old_file, new_file)
+                                    if os.path.exists(old_file):
+                                        os.remove(old_file)
+                                if os.path.exists(f"Oraxen/pack/textures/{get_namespace}/armor/{nv}_leggings.png"):
+                                    old_file_2 = os.path.join(f"Oraxen/pack/textures/{get_namespace}/armor", f"{nv}_leggings.png")
+                                    new_file_2 = os.path.join(f"Oraxen/pack/textures/{get_namespace}/armor", f"{nv}_armor_layer_2.png")
+                                    if not os.path.exists(f"Oraxen/pack/textures/{get_namespace}/armor/{nv}_armor_layer_2.png"):
+                                        os.rename(old_file_2, new_file_2)
+                                    if os.path.exists(old_file_2):
+                                        os.remove(old_file_2)
+                                                                
+                                # set armor 128x32 config
+                                if not os.path.exists(f"Oraxen/settings.yml"):
+                                    shutil.copy(f"Oraxen_settings.yml", f"Oraxen/settings.yml")
                             
                 with open(r'Oraxen\\items\\'+get_file, 'w') as file:
                     documents = yaml.dump(documents['items'], file, Dumper=YmlDumper, default_flow_style=False)
